(impl) c win file scanning - single threaded

+++ c/scanner/process_files.h
@@ -1,6 +1,6 @@
 #ifndef _PROCESS_FILES_H_
 #define _PROCESS_FILES_H_
 
-int ProcessFiles(const char * pathname);
+long ProcessFiles(const char * pathname);
 
 #endif
